+++ Src/Private/Get-AbrADSite.ps1
@@ -41,8 +41,8 @@ function Get-AbrADSite {
 
                             if ($Graph) {
                                 $BestAspectRatio = Get-DiaBestImageAspectRatio -GraphObj $Graph -MaxWidth 600
-                                Section -Style Heading4 $reportTranslate.GetAbrADSite.SiteInventoryDiagram {
-                                    Image -Base64 $Graph -Text $reportTranslate.GetAbrADSite.SiteInventoryDiagram -Width $BestAspectRatio.Width -Height $BestAspectRatio.Height -Align Center
+                                Section -Style Heading4 $reportTranslate.GetAbrADSite.ReplicationDiagram {
+                                    Image -Base64 $Graph -Text $reportTranslate.GetAbrADSite.ReplicationDiagram -Width $BestAspectRatio.Width -Height $BestAspectRatio.Height -Align Center
                                 }
                                 BlankLine -Count 2
                             }