+++ Src/Private/Get-AbrDomainSection.ps1
@@ -79,7 +79,7 @@ function Get-AbrDomainSection {
                                                     if (Get-DCWinRMState -ComputerName $DC -DCStatus ([ref]$DCStatus)) {
                                                         Get-AbrADDCRoleFeature -DC $DC
                                                     } else {
-                                                Write-PScriboMessage -IsWarning -Message ($reportTranslate.GetAbrDomainSection.UnableToConnect -f $DC, $DomainInfo.DNSRoot)
+                                                        Write-PScriboMessage -IsWarning -Message ($reportTranslate.GetAbrDomainSection.UnableToConnect -f $DC, $DomainInfo.DNSRoot)
                                                     }
                                                 }
                                                 if ($RolesObj) {